when entering esc menu the game pauses
