--- Assets/Dungeon/Scripts/Procedural Dungeon Generation/EscMenu/EscConf.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8e71c6ab03c871947b5a41db44ebe247
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
