+++ Assets/Dungeon/Scripts/Procedural Dungeon Generation/EscMenu/EscMenu.cs
@@ -1,6 +1,7 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.SceneManagement;
 
 public class EscMenu : MonoBehaviour
 {
@@ -9,11 +10,14 @@ public class EscMenu : MonoBehaviour
     public GameObject resumeBtn;
     public GameObject exitBtn;
     public GameObject escConfPanel;
+
+    public bool isPaused;
     // Start is called before the first frame update
     void Start()
     {
         gearBtn.SetActive(true);
         escPanel.SetActive(false);
+        escConfPanel.SetActive(false);
     }
 
     // Update is called once per frame
@@ -21,37 +25,66 @@ public class EscMenu : MonoBehaviour
     {
         if (!escPanel.activeInHierarchy)
         {
-            if (Input.GetKeyUp(KeyCode.Escape))
+            if (Input.GetKeyUp(KeyCode.Escape) && !isPaused)
             {
-                escPanel.SetActive(true);
-                gearBtn.SetActive(false);
+                PauseGame();
             }
         }
         else
         {
-            if (Input.GetKeyUp(KeyCode.Escape))
+            if (Input.GetKeyUp(KeyCode.Escape) && isPaused)
             {
-                escPanel.SetActive(false);
-                gearBtn.SetActive(true);
+                ResumeGame();
             }
         }
     }
 
     public void GearButton()
     {
-        escPanel.SetActive(true);
-        gearBtn.SetActive(false);
+        PauseGame();
     }
 
     public void ResumeBtn()
     {
-        escPanel.SetActive(false);
-        gearBtn.SetActive(true);
+        ResumeGame();
     }
 
     public void ExitBtn()
     {
         escPanel.SetActive(false);
         escConfPanel.SetActive(true);
+    }
+
+    public void PauseGame()
+    {
+        escPanel.SetActive(true);
+        gearBtn.SetActive(false);
+        Time.timeScale = 0f;
+        isPaused = true;
+    }
+
+    public void ResumeGame()
+    {
+        escPanel.SetActive(false);
+        gearBtn.SetActive(true);
+        Time.timeScale = 1f;
+        isPaused = false;
+    }
+
+    public void LoadFarm()
+    {
+        SceneManager.LoadScene("Farm");
+    }
+
+    public void Confirm()
+    {
+        LoadFarm();
+    }
+    public void Refuse()
+    {
+        escConfPanel.SetActive(false);
+        gearBtn.SetActive(true);
+        Time.timeScale = 1f;
+        isPaused = false;
     }
 }
