+++ ProjectSettings/TagManager.asset
@@ -45,7 +45,7 @@ TagManager:
   - 
   - 
   - 
-  - 
+  - EscMenu
   - 
   - 
   m_SortingLayers:
